+++ mbgl/map.go
@@ -12,7 +12,7 @@ type Map struct {
 func NewMap(
 	renderer RendererFrontend,
 	size Size,
-	pixelRatio float32,
+	pixelRatio float64,
 	source FileSource,
 	scheduler Scheduler,
 	mapMode MapMode,
@@ -43,15 +43,15 @@ func (m *Map) GetStyle() *Style {
 	return &style
 }
 
-func (m *Map) SetLatLngZoom(latLng *LatLng, zoom float32) {
+func (m *Map) SetLatLngZoom(latLng *LatLng, zoom float64) {
 	C.mbgl_map_set_lat_lng_zoom(C.MbglMap(m.cptr), C.MbglLatLng(latLng.cptr), C.double(zoom))
 }
 
-func (m *Map) SetBearing(degrees float32) {
+func (m *Map) SetBearing(degrees float64) {
 	C.mbgl_map_set_bearing(C.MbglMap(m.cptr), C.double(degrees))
 }
 
-func (m *Map) SetPitch(pitch float32) {
+func (m *Map) SetPitch(pitch float64) {
 	C.mbgl_map_set_pitch(C.MbglMap(m.cptr), C.double(pitch))
 }
